backend/requirements.txt
flask==2.2.5
flask-cors==4.0.0
python-dotenv==1.0.0
flasgger==0.9.5
fastapi~=0.124.4
uvicorn[standard]
numpy>=1.24.0
sentence-transformers>=2.2.0
tqdm>=4.65.0
torch>=2.0.0
pytest>=7.0.0
transformers~=4.45.2
requests~=2.32.3
Pillow~=10.1.0
pydantic~=2.12.5